renaming all react table components
