+++ src/constants/index.js
@@ -14,7 +14,5 @@ export const TICKET_STATUS_OPTIONS = {
 };
 
 export const mobileWidth = '600px';
-
 export const tabletWidth = '960px';
-
 export const desktopWidth = '1200px';
