+++ src/views/managers/managersTableComponents.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 import { Link } from "react-router-dom";
-
-
 export const columns = [
   {
     dataField: "fullName",
