+++ src/views/tenants/tenants.jsx
@@ -12,7 +12,7 @@ import { faChevronRight, faChevronDown } from '@fortawesome/free-solid-svg-icons
 import Search from "../../components/Search/index";
 import { ShowHideSwitch } from '../../components/ShowHideSwitch';
 import Modal from '../../components/Modal';
-import { columns, mobileColumns } from './tenantsFormStructure';
+import { columns, mobileColumns } from './tenantsTableComponents';
 import { mobileWidth } from '../../constants/index.js';
 import { useMediaQueries } from '@react-hook/media-query';
 import './tenants.scss';
